refactor(Utils): /utils.ts -> /utils/utils.ts
